refactor: move getRunningTime to util
