+++ front/src/utils/util.ts
@@ -2,4 +2,12 @@ const winRate = (totalWin: number, totalLose: number) => {
   return Math.floor((totalWin / (totalWin + totalLose)) * 100);
 };
 
-export { winRate };
+const getRunningTime = (isoStartTime: Date): string => {
+  const currentTime: Date = new Date();
+  const rawGameTime = currentTime.getTime() - isoStartTime.getTime();
+  const min = String(Math.floor(rawGameTime / 1000 / 60));
+  const sec = String(Math.floor(rawGameTime / 1000) - Number(min) * 60);
+  return `${min}:${sec}`;
+};
+
+export { winRate, getRunningTime };
